+++ prody/__init__.py
@@ -60,12 +60,10 @@ __all__ = ['checkUpdates', 'confProDy', 'startLogfile', 'closeLogfile', 'plog']
 
 from . import utilities
 from .utilities import *
-from .utilities import PackageLogger, PackageSettings
+from .utilities import LOGGER, PackageSettings
 from .utilities import getPackagePath, joinRepr, tabulate
 __all__.extend(utilities.__all__)
 __all__.append('utilities')
-
-LOGGER = PackageLogger('.prody')
 
 SETTINGS = PackageSettings('prody', logger=LOGGER)
 SETTINGS.load()
