+++ prody/measure/measure.py
@@ -813,5 +813,6 @@ def calcInertiaTensor(coords):
 def calcPrincAxes(coords, turbo=True):
     """Calculate principal axes from coords"""
     M = calcInertiaTensor(coords)
-    _, vectors = solveEig(M, 3)
+    _, vectors = solveEig(M, 3, zeros=True, turbo=turbo, 
+                          warn_zeros=False, reverse=True)
     return vectors
